renamed the header file for TemplatedSecondaryVertexTagInfo so it matches the class name
