+++ DataFormats/BTauReco/interface/CandSecondaryVertexTagInfo.h
@@ -1,0 +1,15 @@
+#ifndef DataFormats_BTauReco_CandSecondaryVertexTagInfo_h
+#define DataFormats_BTauReco_CandSecondaryVertexTagInfo_h
+
+#include "DataFormats/BTauReco/interface/RefMacros.h"
+#include "DataFormats/BTauReco/interface/TemplatedSecondaryVertexTagInfo.h"
+
+namespace reco {
+
+typedef reco::TemplatedSecondaryVertexTagInfo<reco::CandIPTagInfo,reco::VertexCompositePtrCandidate> CandSecondaryVertexTagInfo;
+
+
+DECLARE_EDM_REFS(CandSecondaryVertexTagInfo)
+
+}
+#endif // DataFormats_BTauReco_CandSecondaryVertexTagInfo_h
